feat(gateway): add jwt and service secret verficiation
An authenticated user sends jwt token. The gateway verifies it. If
verified, it attaches 'X-User-Id' and 'X-User-Role' headers to services.
A service does send jwt to gateway. Services uses 'X-Service-Secret'
header to prove their identity. If it's valid, gateway respects
'X-User-Id' and 'X-User-Role' as-is, and convey them to other services.